ImagesCarousel moved in components foulder
- deleted also a few console.log
