+++ src/components/index.js
@@ -6,6 +6,7 @@ export * from './DiagonalGradient';
 export * from './HtmlView';
 export * from './Icon';
 export * from './Image';
+export * from './ImagesCarousel';
 export * from './Link';
 export * from './Logo';
 export * from './Text';
