+++ src/components/screens/EventRecord.js
@@ -13,7 +13,7 @@ import { PriceCard } from './PriceCard';
 import { InfoCard } from './InfoCard';
 import { OperatingCompanyInfo } from './OperatingCompanyInfo';
 import { OpeningTimesCard } from './OpeningTimesCard';
-import { ImagesCarousel } from './home/ImagesCarousel';
+import { ImagesCarousel } from '../ImagesCarousel';
 
 /* eslint-disable complexity */
 /* NOTE: we need to check a lot for presence, so this is that complex */
@@ -34,7 +34,6 @@ export const EventRecord = ({ data }) => {
 
   const logo = dataProvider && dataProvider.logo && dataProvider.logo.url;
   let images = [];
-  console.log(mediaContents.contentType);
 
   !!mediaContents &&
     !!mediaContents.length &&
