+++ src/components/screens/index.js
@@ -1,5 +1,4 @@
 export * from './home/Service';
-export * from './home/ImagesCarousel';
 export * from './EventRecord';
 export * from './InfoCard';
 export * from './PointOfInterest';
